MVVM일부 추가
MVVM일부 추가(AppstoreResultTableViewModel, MainViewControllerModel)

MainViewControllerModel은 Controller상에 viewModel로 이동시킬만한 액션이 보이지 않는다고 판단 되어 일단 사용되지 않은 상태.
